requirements.txt
langchain==1.2.15
langchain-openai==1.2.1
python-dotenv==1.2.2
langgraph==1.1.9
langchain-mcp-adapters==0.2.2
fastmcp==3.2.4
IPython==9.13.0
deepagents==0.5.3
tavily-python==0.7.23
pydantic==2.12.5
typing_extensions==4.15.0
langchain-modal==0.0.3
langfuse==3.14.1
fastapi==0.129.0
uvicorn==0.40.0
sqlalchemy==2.0.46
modal==1.3.2
requests==2.33.0
mcp==1.26.0
google-adk==1.31.1
a2a-sdk==1.0.2
geopy==2.4.1
httpx==0.28.1
click==8.3.1
gradio==6.13.0